// src/app/modules/academicSemester/academicSemester.interface.ts
/* eslint-disable no-unused-vars */
import { Model } from "mongoose";

export type TMonths =
  | "January"
  | "February"
  | "March"
  | "April"
  | "May"
  | "June"
  | "July"
  | "August"
  | "September"
  | "October"
  | "November"
  | "December";

export type TAcademicSemesterName = "Autumn" | "Summer" | "Fall";

export type TAcademicSemesterCode = "01" | "02" | "03";

export interface IAcademicSemester {
  _id: string;
  name: TAcademicSemesterName;
  year: string;
  code: TAcademicSemesterCode;
  startMonth: TMonths;
  endMonth: TMonths;
  isDeleted: boolean;
}

export interface IAcademicSemesterCodeMapper {
  [key: string]: string;
}

export interface IAcademicSemesterModel extends Model<IAcademicSemester> {
  findOneOrThrowError: (id: string) => Promise<IAcademicSemester>;
}